an attempt to make a more sophisticated reprensentation
it turns out to be a stupid attempt, but listen more.

So, since we're going to drop the signedness bit some ideas about how
we can reuse it arose. One of the ideas, was that we can use a prefix
notation and represent common widths using 3 bits, and 14 bits for the
arbitrary legnth. This would allow us to save space in the immediate
word to store more payload. Those common values were 1,8,32, and 64.

The idea proved to be dumb, because we were saving space for widths
which fit into the immediate even when we use 15 bits for metadata
(15+32=47). So it only makes sence for 64 bit words, and since we have
only one special sized value it would be enough to use only 1 bit (the
signedness bit) to tell that we're dealing with 64 bit value. So we
can use the rest 62 bits of the OCaml immediate for the payload.