adds available expression optimization
It is an experiment that implements let-driven available-expression
optimization.

The implementation is simple,

1) every time we see a new let-binding, if the expression is already
   bound, we do not create any new binding, but produce the body
   directly, with the variable name substituted if needed.

2) when a let binding created we are making it available in the body,
   by pushing it to the state. We map expressions to variables, and
   (1) ensures that each expression is mapped to at least one
   variable.

3) when we create an expression, we see if it is available already,
   and if its, then we spill out the corresponding variable instead.

The experiment didn't show a significant reduce in the size of
generated expressions, mostly because our code already produces rather
optimized expressiosn. In other words, we manually use a lot of lets,
and rarely, if at all, do double letting, or forget to use the already
let-bound expression. So, we are considering is whether this work is
worthwhile right now.

Note, currently the implementation is unfit, as the substitution is
not bound-free, and a few other shortcommings, which should affect
the correctnes of the system in its current stage, but in general
are not safe.