a (failed) experiment with a new implementation of records
Instead of using Univ_map we can use arrays to represent records.
The initial assumptions were
1) we do more collect than provide, and while provide would be a
   little bit slower this would be ammortized with a very fast
   collect
2) most of the classes are small
3) the number of slots doesn't change during runtime

Assumption 3 doesn't really hold as libraries (and plugins) can easily
define values on the toplevel before other libraries (and plugins) are
loaded, therefore the implementation still has to consider values with
different number of slots.

But the main problem is that assumption 2 doesn't hold at all, because
we have a bloated class Theory.Program for which we've added dozens if
not hundreds of slots. Therefore, every time we create an empty
program it creates an huge array and then copies it, and so on.

As a result we ended up 2 times slower. Not that bad, given that every
program value had more than 50 slots.