projects ELF section flags to the is-valid attribute
Proposal
========

If a section is non-executable, then mark all addresses that belong to
this section as not-valid, so that the disassembler won't even try to
disassemble them.

Context
=======

ELF files have two different views on their layout - segments and
sections. BAP loaders are using the segment view, in which there is
a big contiguous loadable and executable segment (usually named "2"),
which contains many sections, including .plt, .text, and, even
.rodata.

The loader, at least on Linux, is mapping the whole segment to memory,
but then assigns different permission attributes based on _sections_,
so despite the fact that .rodata belongs to an executable section it
is still marked as non-executable by the kernel.

Therefore, we can end up disassembling the rodata section (and some
other non-executable sections of executable segments). We can reach
those sections either via byteweight's false positives or through our
new speculative disassembler.

Usually, the disassembler is able to rebuke most if not all functions
in the non-executable code sections by proving that all chains of
executions terminate with data. This, however takes time, and is still
not guaranteed.

Therefore, it sounds reasonable to use the available information, and
do not disassemble non-executable sections. Thus we can disassemble
faster and our analyses won't spend time analyzing functions that will
never be executed.

Problems
========

The overall bap infrastructure is build on the segment view, so other
components are not aware of existence of sections and their flags and
strongly believe that if a segment is executable, then it is
executable. For example, if we will enable this proposal, then
the information provided by the byteweight will conflict, since
byteweight will scan the whole memory region. Of course, we can fix
it, but how many more components do we need to fix? Probably, we
should fix the problem on the loader level, which shouldn't generate
incorrectly labeled memory regions.

So far, the proposed solution doesn't give any improvements. At least
in the current state. The disassembling time is not improved (in fact
it is even a bit slower, since it is faster for a disassembler to
prove that data is invalid, than quering interval trees of many
sections), second the disassembly is not changing, at least for
typical files, like /bin/ls, /bin/bash, etc.

Suggestions
===========

We should fix our LLVM loader (and probably other loaders).