enables the upper limit if the limiter is used
It was quite unintuitive that to limit the maximum path you need to
enable the limiter and then set the limit to some finite number as
by default it is unlimited.

This were making sense, however, in the old days, when we had only one
system and all components were automatically included into
it. Therefore, to prevent the limiter from jumping in into the old
style bap:legacy-main and to enable seamless backward compatibility,
we did the following:

1) When either --primus-limit-length or --primus-limit-max-depth are
specified we add the component to the legacy-main system.
2) We also add this component when the promiscuous mode is enabled.