implements support for modern C runtime
In the modern C runtime the __libc_start_main function is called not
via the PLT indirection, but through a .got.plt (or .got.plt, can't
remember which is which) table which is zeroed by default (and it is
expected that the loaded will fill in the values from the relocation
table).

Our existing support for C runtime was searching the libc_start_main
location from the entry point - the first function called by the entry
point is assumed to be th libc_start_main, and the first argument of
that function is assumed to the main function.

The problem is that since there is no longer the plt entry for
libc_start_main we can't really link its stub, so the fix is simple,
just insert the synthetic subroutine for it. And everything else works
more or less fine.

This commit also adds the stack canary provision code, for systems
that require it. It allocates the space for the canary and initializes
the FS_BASE variable.

Sine FS_BASE is not added to the Primus environment by default we also
modify the x86-primus-loader plugin and initialize all registers from
x86 family (including the ymms registers).