fixes parent refutation and keeps parents inputs
First of all, parent task constraints weren't refuted and were added
as if we were always taking the yes-branch. We hotfix it to the
negation of the condition, when it was refuted, but in general we
should rething this parent constraints.

Next, we didn't consider the parent inputs so while the model included
them they were never set, since the child thread wasn't considering
them as inputs (as they were already set).