provides the semantics to the knowledge base
A lifter is a piece of code that promises
`Theory.Program.Semantics.slot`.

I had to remove the address, because `bap mc` doesn't provide the
address for instruction (sic). Not a big deal, as we have the address
in the memory chunk. We will update the upstream later, but for now,
let's just not ask for the address (note this is only specific to `bap
mc`, for normal `bap` the address is provided).