enables interworking in the disassembler driver
What is interworking
--------------------

Interworking is a feature of some architectures that enables mixing
several instruction sets in the same compilation unit. Example, arm
and thumb interworking that this branch is trying to add.

What is done
-------------

1. we add the switch primitive to the basic interface that changes the
dissassembler in the current disassembling state. It is a bold move
and can have conseqeuences, should be carefully reviewed

2. attributes each destination in the disassembler driver state with
the architecture and calls switch every time we are going to
disassemble the next chunk of memory.

What is to be done
------------------

Right now the instruction ISA is equal to its module ISA, which not
true for architectures that have interworking, e.g., ARMv5T and above
units contain two ISA, A32 and T32. We need to disable the rule that
establishes equality between units and instruction architecture, at
least of armv5 and above. (Need to double check is armv5 as in BAP is
included).

Next, the arm lifter shall provide a promise to compute
destinations (which itself will require destinations, because we don't
really want to compute them) and provide the destination architecture,
based on the source encoding. We can safely examine any representation
of the instruction since it is already will be lifted by that moment.