implements stub discover for PowerPC stub entries
Unlike other elves, in PowerPC the plt stubs are residing directly in
the .text section and we can't find any pointers to them in the elf
meta information. The good news is that many of them start from the
same sequence of bytes, which we employ. False positives are possible,
but they should have minimal impact, as we will just lauch our stub
recongnition algorithm in the relocatable symbolizer, which with the
very high propability will disregard the following lines as a valid
stub (it has to evaluate to a known relocation), and if it not, then
it will only give a name to that chunk of bytes, not identify it is a
function start or anything like this.