KB semantics tweaks: restarts promises if nothing was computed
Our current semantics is to run promises only once per object's
slot. However, if the promises weren't able to provide any
information, because of the lack of knowledge, then we will not rerun
them later when the information become available.

That branch is a crooked experiment in which we do not mark the slot
as computed if it doesn't have any value stored, even when the
promises were run.