interns all names
Now the KB.Name.t is represented as an int63 of its hash. It is now
allowed to have two different names with the same name, and in case
of a collision the error is reported. Instead of using the
`hash_variant` function from OCaml we rely on our own hash function
that uses int63 and is thus both portable between 32/64/js runtimes
and has much lower chance of a collision.

We now use Names as the class identifiers, which enables portability
of the knowledge base between different versions of BAP (before that
the class id was stored in the knowledge base, but was really
dependent on the module initialization order).